Added Base RobotBuilder Classes
I finished the KVStore interface and the property store class. Component
Builder is still to be done.
