+++ src/main/java/org/usfirst/frc/team4213/metallib/util/KVStore.java
@@ -1,0 +1,8 @@
+package org.usfirst.frc.team4213.metallib.util;
+
+public interface KVStore {
+	public double getDouble(String key);
+	public int getInt(String key);
+	public String getString(String key);
+	public boolean getBool(String key);
+}
